requirements.txt
joblib==1.4.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqt_toast_notification==1.3.2
pyqtgraph==0.13.7
Requests==2.32.3
scikit_learn==1.6.1
setuptools==58.0.4
ta==0.11.0
tensorflow==2.19.0
